[Bugfix] Fall back to outlines for grammars that can't convert to GBNF
